requirements_dev.txt
requests
aioredis
aiohttp
gunicorn
jinja2
pydantic
lxml
bs4
python-dotenv

pytest
pytest-cov
coveralls